requirements.txt
aiogram>=3.4.0
fastapi>=0.111.0
pystray>=0.19.0
Pillow>=10.0.0
uvicorn>=0.29.0
python-multipart>=0.0.9
openai-whisper>=20231117
ffmpeg-python>=0.2.0
SQLAlchemy>=2.0.29
python-dotenv>=1.0.1
google-genai>=0.3.0
openai>=1.14.0
